fix: 백그라운드 마우스 효과 컴포넌트 생성
Layout.tsx 화면 처리가 아닌 컴포넌트를 생성하여 화면 ReactNode를 하위로 뿌려지는 식으로 설계 변경
